# perf(array): optimize array compact with Bitmap::iter_ones()